+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.1.0] - April 12, 2020
+
+* Big fix: CheckboxIconFormField.enabled was not working
+* Default colors can be set by providing `context` properly
+* Update README.md
+
+
 ## [0.0.1] - April 12, 2020
 
 * Add CheckboxListTileFormField